Separate out the type and value into separate interfaces.
State on implementations was getting confusing, when it could be
modified/used, how it would be instantiated, etc. By separating out
types and values into their own interfaces, the boundaries become much
clearer and the answers seem more intuitive. We no longer need
factories--the type can be the factory for the values.